backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=9.0.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
    "gradio-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.45",
]
